Extract generateTeamNames and generateRoleGroups to shared helpers
